Front page documentation additions
The term 'stream' in programming can means lots of different things and
'activity' doesn't narror it down much so we try to clarify the exact
purpose of the project.

Provide explicit example of Github activity stream as a common usage and
one that user's will probably be familiar with.

Explicity mention support of Django models to make it a little clearer
for beginners.

Add 'Github repository' to activity stream example to clarify a little
as without it an 'activity-stream' is quite an abstract concept which is
pretty general in the context of the project but knowing that we're
actutally talking about the repository or project as a real life object
is useful.