Reduce & move warning and clarify Manager role
When I first read this page my attention was drawn entirely to the
warning, missing the very important paragraph above it. The warning is
only relevant to older more familiar users who would be looking for an
explanation of behaviour change and possibly find it in the change log
anyway.

I've change it to a note and moved it down to push focus towards the
main text.

Provided some extra sentences about the Manager. Many Django users won't
be familiar with the term 'Manager' as you can do alot with just the
MyModel.objects manager that you can by default without thinking of it
as anything other than 'that thing I used for queries'.

Clarify that you can only have one activity-stream manager per app with
the current configuration set up.