Expand on generating actions documentation
Link to the Django signals documentation.

Break part the second example and be a little more explicit about what
'group' and 'comment' might be in this example.

Describe a little of how to think about the generated activities and
where they are stored and how they are queried. I feel it is important
to help users build a picture of the usage and limitations of the
system.