Check for exceptions when fetching object by content type
As the get_object_for_this_type function states,
"The ObjectNotExist exception, if thrown, will not be caught,
 so code that calls this method should catch it."

This means that if we do not have an object with id 89 but do have
a content type with id 12 and call
GET activity/feed/12/89
the app would throw an ObjectDoesNotExist exception, instead
of nicely returning a 404 error as in the case of a non matching
content type.

The returned error message is the same as what you get from
get_object_or_404().