Added explicit 'on_delete' arg to ForeignKey fields
Beginning in Django 2.0, the 'on_delete' argument will be required for any
ForeignKey model fields.  The default behavior in Django is to CASCADE delete
a Model when its related field is removed, so I added this definition to the
existing Model fields:

- Follow.user
- Follow.content_type
- Action.actor_content_type
- Action.target_content_type
- Action.action_object_content_type

These changes should be backwards-compatible with any existing installation.
I also added these definitions to the migrations files--though not the
south_migrations (since it's unlikely South would be used in a Django 2.0
installation at this point).

I did this mainly to get rid of the many 'RemovedInDjango20Warning'
deprecation notices in my Django 1.9 installation, but will allow the module
to be compatible with future versions of Django.