Use getattr directly to get username field name
The other method does not work properly in this case since
it will return a function and that function is passed as a key to
a  `get_object_or_404` in `user()` later on. That would not work
since a string has to be passed as the key for a field when querying.