Remove the combination of django21 and sqlite, as travis can't satisfy it.
The problem here is that the Travis tests are based on an image using
Ubuntu Trusty (14.04) which is limited to sqlite3 Version 3.8.2 or below.
There is no easy way to upgrade, it may be simpler (and a better idea) to
upgrade to a newer image, e.g. Xenial (16.04). This is only problematic
with Django 2.1+.