Try to fix recent test failures (#426)
* Fix p37-django21-mysql environment not requiring mysqlclient.

* Fix error stating django.contrib.messages must be in INSTALLED_APPS.

* Add py37 to travis file so tests may actually run against it.

So far they seem to be skipped as the interpreter is missing.

* Remove the combination of django21 and sqlite, as travis can't satisfy it.

The problem here is that the Travis tests are based on an image using
Ubuntu Trusty (14.04) which is limited to sqlite3 Version 3.8.2 or below.
There is no easy way to upgrade, it may be simpler (and a better idea) to
upgrade to a newer image, e.g. Xenial (16.04). This is only problematic
with Django 2.1+.

* Remove py37 for now, the trusty image does not support it.

* Remove combination of sqlite and Django 2.1 also from travis file.

I had removed it from the tox file, resulting in this being build in travis
but not doing anything. Better to not even build it.

* Try fixing py34 environments by pinning mysqlclient version.

That version used to work during the last PR on master, so let's see if
it changes something. The 'mysqlclient>=1.3.13' will currently install
version 1.4.2, which may be incompatible with py34.

* Avoid defining mysqlclient dep twice. Limit to below 1.4 to try and fix py34.

* Avoid defining mysqlclient dep twice. Limit to below 1.4 to try and fix py34.

* Order env vars in a way their grouping is more obvious.